intervals

+++ include/includes.h
@@ -29,7 +29,7 @@ extern std::atomic<log_mode_t> logMode;
 extern std::atomic<bool> logAppend;
 extern std::atomic<log_cmd_t> logCmd;
 
-extern ulong firstFix;
+extern uint32_t firstFix;
 extern gps_state_ctx_t gps_state;
 extern logfileW logfile;
 extern const size_t &fsTotalBytes;
